backend/session/requirements.txt
pydantic==2.9.2
psycopg2-binary==2.9.9